Part 2 of 2 of upload
The last set of files from the old repo. This version from memory is quite broken as most Persona Fusion types do not work. If the Project does not seem viable, I will abandon it in favour for new projects - as this one was started when I had a lot less experience. Sadly, this abandonment looks more likely; as the code is of relatively poor quality and effectively needs rewriting from scratch.  However, I will leave the files here as an archive.
(Original date of project start: Feb 2019 at the latest)